+++ pkg/ruler/compat_test.go
@@ -226,12 +226,13 @@ func TestMetricsQueryFuncErrors(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			queries := prometheus.NewCounter(prometheus.CounterOpts{})
 			failures := prometheus.NewCounter(prometheus.CounterOpts{})
+			queryTime := prometheus.NewCounterVec(prometheus.CounterOpts{}, []string{"user"})
 
 			mockFunc := func(ctx context.Context, q string, t time.Time) (promql.Vector, error) {
 				return promql.Vector{}, tc.returnedError
 			}
 
-			qf := metricsQueryFunc(mockFunc, queries, failures)
+			qf := metricsQueryFunc(mockFunc, queries, failures, queryTime, "user")
 
 			_, err := qf(context.Background(), "test", time.Now())
 			require.Equal(t, tc.returnedError, err)